Refactored JSON file handling and updated package structure
