+++ app/src/main/AndroidManifest.xml
@@ -14,9 +14,9 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AppCompat.Light.DarkActionBar"
         tools:targetApi="31"
-        android:name=".YoutubeApplication">
+        android:name="io.github.romantsisyk.youtube.YoutubeApplication">
         <activity
-            android:name=".presentation.video_list.VideoListActivity"
+            android:name="io.github.romantsisyk.youtube.presentation.video_list.VideoListActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
@@ -25,7 +25,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".presentation.video_detail.SingleVideoActivity"
+        <activity android:name="io.github.romantsisyk.youtube.presentation.video_detail.SingleVideoActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.SEND" />
