--- app/src/main/java/com/roman_tsisyk/youtube/YoutubeApplication.kt
@@ -1,7 +1,0 @@
-package com.roman_tsisyk.youtube
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class YoutubeApplication : Application()
