--- app/src/main/java/com/roman_tsisyk/youtube/data/model/GithubVideoResponse.kt
@@ -1,3 +1,0 @@
-package com.roman_tsisyk.youtube.data.model
-
-data class GithubVideoResponse(val videoIds: List<String>)
