--- app/src/main/java/com/roman_tsisyk/youtube/data/source/local/LocalVideoDataSource.kt
@@ -1,4 +1,0 @@
-package com.roman_tsisyk.youtube.data.source.local
-
-class LocalVideoDataSource {
-}
