--- app/src/main/java/com/roman_tsisyk/youtube/domain/model/Video.kt
@@ -1,3 +1,0 @@
-package com.roman_tsisyk.youtube.domain.model
-
-data class Video(val id: String)
