--- app/src/main/java/com/roman_tsisyk/youtube/domain/repository/VideoRepository.kt
@@ -1,8 +1,0 @@
-package com.roman_tsisyk.youtube.domain.repository
-
-import com.roman_tsisyk.youtube.domain.model.Video
-
-interface VideoRepository {
-    suspend fun fetchVideos(): List<Video>
-    suspend fun getVideoDetail(videoId: String): Video?
-}
