--- app/src/main/java/com/roman_tsisyk/youtube/domain/usecase/FetchVideosUseCase.kt
@@ -1,10 +1,0 @@
-package com.roman_tsisyk.youtube.domain.usecase
-
-import com.roman_tsisyk.youtube.domain.model.Video
-import com.roman_tsisyk.youtube.domain.repository.VideoRepository
-
-class FetchVideosUseCase(private val repository: VideoRepository) {
-    suspend fun execute(): List<Video> {
-        return repository.fetchVideos()
-    }
-}
