--- app/src/main/java/com/roman_tsisyk/youtube/domain/usecase/GetVideoDetailUseCase.kt
@@ -1,10 +1,0 @@
-package com.roman_tsisyk.youtube.domain.usecase
-
-import com.roman_tsisyk.youtube.domain.model.Video
-import com.roman_tsisyk.youtube.domain.repository.VideoRepository
-
-class GetVideoDetailUseCase(private val repository: VideoRepository) {
-    suspend fun execute(videoId: String): Video? {
-        return repository.getVideoDetail(videoId)
-    }
-}
