--- app/src/main/java/com/roman_tsisyk/youtube/presentation/splash/SplashActivity.kt
@@ -1,19 +1,0 @@
-package com.roman_tsisyk.youtube.presentation.splash
-
-import android.annotation.SuppressLint
-import android.app.Activity
-import android.content.Intent
-import android.os.Bundle
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.roman_tsisyk.youtube.presentation.video_list.VideoListActivity
-
-@SuppressLint("CustomSplashScreen")
-class SplashActivity : Activity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        installSplashScreen()
-        startActivity(Intent(this, VideoListActivity::class.java))
-        finish()
-    }
-}
