+++ app/src/main/java/io/github/romantsisyk/youtube/YoutubeApplication.kt
@@ -1,0 +1,7 @@
+package io.github.romantsisyk.youtube
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class YoutubeApplication : Application()
