+++ app/src/main/java/io/github/romantsisyk/youtube/data/model/GithubVideoResponse.kt
@@ -1,0 +1,3 @@
+package io.github.romantsisyk.youtube.data.model
+
+data class GithubVideoResponse(val videoIds: List<String>)
