+++ app/src/main/java/io/github/romantsisyk/youtube/data/source/local/LocalVideoDataSource.kt
@@ -1,0 +1,4 @@
+package io.github.romantsisyk.youtube.data.source.local
+
+class LocalVideoDataSource {
+}
