+++ app/src/main/java/io/github/romantsisyk/youtube/domain/model/Video.kt
@@ -1,0 +1,3 @@
+package io.github.romantsisyk.youtube.domain.model
+
+data class Video(val id: String)
