+++ app/src/main/java/io/github/romantsisyk/youtube/domain/repository/VideoRepository.kt
@@ -1,0 +1,8 @@
+package io.github.romantsisyk.youtube.domain.repository
+
+import io.github.romantsisyk.youtube.domain.model.Video
+
+interface VideoRepository {
+    suspend fun fetchVideos(): List<Video>
+    suspend fun getVideoDetail(videoId: String): Video?
+}
